web/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.16
django-cors-headers==3.11.0
django-admin-display==1.3.0
django-admin-ip-restrictor==2.2.0
django-ipware==3.0.7
graphql-core==3.2.3
idna==3.4
pandas==1.5.0
packaging==21.3
Pillow==9.2.0
Pygments==2.13.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
# strawberry-django-jwt==0.2.2
strawberry-django-auth==0.3.6.0
strawberry-django-plus==1.27
strawberry-graphql==0.132.1
strawberry-graphql-django==0.6
typing_extensions==4.3.0
urllib3==1.26.12
uWSGI==2.0.20
cryptography
# django-property-filter
mysqlclient==2.1.0
# Opencv-python==4.9.0.80
# opencv-python-headless==4.9.0.80